% Generated by roxygen2 (4.0.0): do not edit by hand

\name{enviroCaR-package}
\alias{enviroCaR-package}
\alias{enviroCaR}
\alias{envirocar.org}
\docType{package}
\title{
enviroCaR-Package
}
\description{
The package provides basic functions to load and analyse measurements from the envirocar server.
}
\details{
\tabular{ll}{
Package: \tab enviroCaR\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2014-03-14\cr
License: \tab GPL (>= 2)\cr
}

%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
Edzer Pebesma
Christoph Stasch
Jan Wirwahn

Maintainer: Christoph Stasch <staschc@uni-muenster.de>
}
\references{
Further material:
Trajectories represented by a lists of lists: http://rpubs.com/edzer/trajectories
Analyzing Envirocar trajectory data with R: http://rpubs.com/edzer/envirocar2
}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
