% Generated by roxygen2 (4.0.0): do not edit by hand

\name{getTrackIDs}
\alias{getTrackIDs}
\title{retrieves the track IDs from the Envirocar server for passed spatial and/or temporal filter}
\usage{
getTrackIDs(serverUrl, bbox, timeInterval, verbose = FALSE)
}
\arguments{
  \item{serverUrl}{base URL of the Envirocar server}

  \item{bbox}{spatial bounding box}

  \item{timeInterval}{interval represented as list of
  POSIXct}

  \item{verbose}{print debug output}
}
\value{
list containing the track IDs for the specified bbox and
time interval, if these are present; otherwise all track
IDs are returned
}
\description{
retrieves the track IDs from the Envirocar server for
passed spatial and/or temporal filter
}

