\name{importEnviroCar}
\alias{importEnviroCar}
\title{Imports the envirocar data}
\usage{
  importEnviroCar(serverUrl, trackIDs, bbox, timeInterval)
}
\arguments{
  \item{serverUrl}{url to server}

  \item{trackIDs}{ids of tracks that should be retrieved}

  \item{bbox}{spatial bounding box}

  \item{timeInterval}{interval}
}
\value{
  Tracks objects for the requested tracks
}
\description{
  Imports the envirocar data
}

