\name{importSingleTrack}
\alias{importSingleTrack}
\title{Imports a single track}
\usage{
importSingleTrack(serverUrl, trackID, verbose = FALSE)
}
\arguments{
  \item{serverUrl}{url to server}

  \item{trackID}{ids of the track that should be retrieved}
}
\value{
Tracks objects for the requested tracks

TODO: unclear how to encode temporalFilter!!
}
\description{
Imports a single track
}

