\name{parseLinkHeaderParam}
\alias{parseLinkHeaderParam}
\title{ugly function for parsing the Link header parameter from the HTTP header, used for the paging mechanism}
\usage{
parseLinkHeaderParam(headerParam)
}
\arguments{
  \item{headerParam}{url to server}
}
\value{
Tracks objects for the requested tracks TODO: unclear how
to encode temporalFilter!!
}
\description{
ugly function for parsing the Link header parameter from
the HTTP header, used for the paging mechanism
}

