\name{parseTrackIDs}
\alias{parseTrackIDs}
\title{function that is used internally for parsing the trackIDs from JSON response}
\usage{
parseTrackIDs(jsonBody)
}
\arguments{
  \item{jsonBody}{that is return for a tracks URL from
  Envirocar server}
}
\value{
list containing the parsed trackIDs
}
\description{
function that is used internally for parsing the trackIDs
from JSON response
}

