\name{envirocaR-package}
\alias{envirocaR-package}
\alias{envirocaR}
\docType{package}
\title{
envirocaR-Package
}
\description{
The package provides basic functions to load and analyse measurements from the envirocar server.
}
\details{
\tabular{ll}{
Package: \tab envirocaR\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2014-02-27\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Edzer Pebesma
Christoph Stasch
Jan Wirwahn
Christopher Stephan

Maintainer: Edzer Pebesma <yourfault@somewhere.net>
}
\references{
Further material:
Trajectories represented by a lists of lists: http://rpubs.com/edzer/trajectories
Analyzing Envirocar trajectory data with R: http://rpubs.com/edzer/envirocar2
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}