\name{importSingleTrack}
\alias{importSingleTrack}
\title{Imports a single track}
\usage{
importSingleTrack(serverUrl, trackID, verbose = FALSE)
}
\arguments{
  \item{serverUrl}{url to server}

  \item{trackID}{ids of the track that should be retrieved}

  \item{verbose}{print debug output}
}
\value{
Tracks objects for the requested tracks
}
\description{
Imports a single track
}

