% Generated by roxygen2 (4.0.0): do not edit by hand

\name{parseLinkHeaderParam}
\alias{parseLinkHeaderParam}
\title{ugly function for parsing the Link header parameter from the HTTP header, used for the paging mechanism}
\usage{
parseLinkHeaderParam(headerParam)
}
\arguments{
  \item{headerParam}{url to server}
}
\value{
Tracks objects for the requested tracks
}
\description{
ugly function for parsing the Link header parameter from
the HTTP header, used for the paging mechanism
}

